changement nom encode/decode + changement type index boucle

+++ encode.c
@@ -27,7 +27,7 @@ int main(int argc, char *argv[]) {
     uint8_t key[16] =  {0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0A, 0x0B, 0x0C, 0x0D, 0x0E, 0x0F};
     keyExpansion(key, extKey);
 
-    for (int i = 0; i < sizeFile; i++) {
+    for (size_t i = 0; i < sizeFile; i++) {
         if(i % 16 == 0 && i != 0){
             uint8_t* tab = malloc(16*sizeof(uint8_t));
             cipher(in, out, extKey);
@@ -38,7 +38,7 @@ int main(int argc, char *argv[]) {
         }
         in[i % 16] = text[i];
         if (i==sizeFile-1){
-            for(int j=(i+1)%16;j<16;j++){
+            for(size_t j=(i+1)%16;j<16;j++){
                 in[j] = 0;
                 printf("%d\n",j);
             }
